Traktor: Fixed compile issue in render context. Shader pipeline generate platform permutation first instead for each technique.
[git-p4: depot-paths = "//depot/traktor/": change = 6144]

+++ code/Render/Context/RenderContext.cpp
@@ -12,8 +12,8 @@ namespace traktor
 T_FORCE_INLINE bool SortOpaquePredicate(const RenderBlock* renderBlock1, const RenderBlock* renderBlock2)
 {
 	const float c_distanceQuantizeRangeInv = 1.0f / 4.0f;
-	float d1 = std::floorf(renderBlock1->distance * c_distanceQuantizeRangeInv);
-	float d2 = std::floorf(renderBlock2->distance * c_distanceQuantizeRangeInv);
+	float d1 = std::floor(renderBlock1->distance * c_distanceQuantizeRangeInv);
+	float d2 = std::floor(renderBlock2->distance * c_distanceQuantizeRangeInv);
 
 	if (d1 < d2)
 		return true;
